adding event Date details in its own component
